client/requirements.txt
# Desktop Client Requirements
# All libraries are FREE and work OFFLINE

# Wake word detection (offline)
vosk>=0.3.45

# Speech-to-Text (offline)
openai-whisper>=20231117

# Text-to-Speech (offline)
pyttsx3>=2.90

# Audio handling
pyaudio>=0.2.13
sounddevice>=0.4.6
numpy>=1.24.0

# Gateway communication
httpx>=0.25.0
websockets>=12.0
requests>=2.31.0
